+++ src/main/java/org/jabref/gui/collab/EntryAddChangeViewModel.java
@@ -46,7 +46,7 @@ class EntryAddChangeViewModel extends DatabaseChangeViewModel {
     @Override
     public Node description() {
         PreviewViewer previewViewer = new PreviewViewer(new BibDatabaseContext(), dialogService, stateManager, themeManager);
-        previewViewer.setLayout(preferencesService.getPreviewPreferences().getCurrentPreviewStyle());
+        previewViewer.setLayout(preferencesService.getPreviewPreferences().getSelectedPreviewLayout());
         previewViewer.setEntry(entry);
         return previewViewer;
     }