+++ src/main/java/org/jabref/gui/preview/CopyCitationAction.java
@@ -15,6 +15,7 @@ import org.jabref.gui.StateManager;
 import org.jabref.gui.actions.ActionHelper;
 import org.jabref.gui.actions.SimpleCommand;
 import org.jabref.gui.util.BackgroundTask;
+import org.jabref.gui.util.TaskExecutor;
 import org.jabref.logic.citationstyle.CitationStyleGenerator;
 import org.jabref.logic.citationstyle.CitationStyleOutputFormat;
 import org.jabref.logic.citationstyle.CitationStylePreviewLayout;
@@ -22,6 +23,7 @@ import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.layout.Layout;
 import org.jabref.logic.layout.LayoutFormatterPreferences;
 import org.jabref.logic.layout.LayoutHelper;
+import org.jabref.logic.layout.TextBasedPreviewLayout;
 import org.jabref.logic.preview.PreviewLayout;
 import org.jabref.logic.util.OS;
 import org.jabref.model.entry.BibEntry;
@@ -38,21 +40,29 @@ public class CopyCitationAction extends SimpleCommand {
     private static final Logger LOGGER = LoggerFactory.getLogger(CopyCitationAction.class);
 
     private final List<BibEntry> selectedEntries;
-    private final StateManager stateManager;
-    private final PreviewLayout style;
-    private final String previewStyle;
+    private final PreviewLayout previewLayout;
+    private final TextBasedPreviewLayout customPreviewLayout;
     private final CitationStyleOutputFormat outputFormat;
+
+    private final StateManager stateManager;
     private final DialogService dialogService;
     private final ClipBoardManager clipBoardManager;
+    private final TaskExecutor taskExecutor;
 
-    public CopyCitationAction(CitationStyleOutputFormat outputFormat, DialogService dialogService, StateManager stateManager, ClipBoardManager clipBoardManager, PreviewPreferences previewPreferences) {
-        this.selectedEntries = stateManager.getSelectedEntries();
-        this.style = previewPreferences.getCurrentPreviewStyle();
-        this.previewStyle = previewPreferences.getPreviewStyle();
+    public CopyCitationAction(CitationStyleOutputFormat outputFormat,
+                              DialogService dialogService,
+                              StateManager stateManager,
+                              ClipBoardManager clipBoardManager,
+                              TaskExecutor taskExecutor,
+                              PreviewPreferences previewPreferences) {
         this.outputFormat = outputFormat;
-        this.clipBoardManager = clipBoardManager;
         this.dialogService = dialogService;
         this.stateManager = stateManager;
+        this.selectedEntries = stateManager.getSelectedEntries();
+        this.clipBoardManager = clipBoardManager;
+        this.taskExecutor = taskExecutor;
+        this.previewLayout = previewPreferences.getSelectedPreviewLayout();
+        this.customPreviewLayout = previewPreferences.getCustomPreviewLayout();
 
         this.executable.bind(ActionHelper.needsEntriesSelected(stateManager));
     }
@@ -62,33 +72,39 @@ public class CopyCitationAction extends SimpleCommand {
         BackgroundTask.wrap(this::generateCitations)
                       .onFailure(ex -> LOGGER.error("Error while copying citations to the clipboard", ex))
                       .onSuccess(this::setClipBoardContent)
-                      .executeWith(Globals.TASK_EXECUTOR);
+                      .executeWith(taskExecutor);
     }
 
     private List<String> generateCitations() throws IOException {
         // This worker stored the style as filename. The CSLAdapter and the CitationStyleCache store the source of the
         // style. Therefore, we extract the style source from the file.
         String styleSource = null;
-        if (style instanceof CitationStylePreviewLayout) {
-            styleSource = ((CitationStylePreviewLayout) style).getSource();
+
+        if (previewLayout instanceof CitationStylePreviewLayout citationStyleLayout) {
+            styleSource = citationStyleLayout.getSource();
         }
+
         if (styleSource != null) {
             return CitationStyleGenerator.generateCitations(selectedEntries, styleSource, outputFormat);
         } else {
-            if (stateManager.getActiveDatabase().isEmpty()) {
-                return Collections.emptyList();
-            }
-
-            StringReader sr = new StringReader(previewStyle.replace("__NEWLINE__", "\n"));
-            LayoutFormatterPreferences layoutFormatterPreferences = Globals.prefs.getLayoutFormatterPreferences(Globals.journalAbbreviationRepository);
-            Layout layout = new LayoutHelper(sr, layoutFormatterPreferences).getLayoutFromText();
-
-            List<String> citations = new ArrayList<>(selectedEntries.size());
-            for (BibEntry entry : selectedEntries) {
-                citations.add(layout.doLayout(entry, stateManager.getActiveDatabase().get().getDatabase()));
-            }
-            return citations;
+            return generateTextBasedPreviewLayoutCitations();
         }
+    }
+
+    private List<String> generateTextBasedPreviewLayoutCitations() throws IOException {
+        if (stateManager.getActiveDatabase().isEmpty()) {
+            return Collections.emptyList();
+        }
+
+        StringReader customLayoutReader = new StringReader(customPreviewLayout.getText().replace("__NEWLINE__", "\n"));
+        LayoutFormatterPreferences layoutFormatterPreferences = Globals.prefs.getLayoutFormatterPreferences(Globals.journalAbbreviationRepository);
+        Layout layout = new LayoutHelper(customLayoutReader, layoutFormatterPreferences).getLayoutFromText();
+
+        List<String> citations = new ArrayList<>(selectedEntries.size());
+        for (BibEntry entry : selectedEntries) {
+            citations.add(layout.doLayout(entry, stateManager.getActiveDatabase().get().getDatabase()));
+        }
+        return citations;
     }
 
     /**
@@ -174,7 +190,7 @@ public class CopyCitationAction extends SimpleCommand {
 
     private void setClipBoardContent(List<String> citations) {
         // if it's not a citation style take care of the preview
-        if (!(style instanceof CitationStylePreviewLayout)) {
+        if (!(previewLayout instanceof CitationStylePreviewLayout)) {
             clipBoardManager.setContent(processPreview(citations));
         } else {
             // if it's generated by a citation style take care of each output format